refactor: Updated src/capture.py (#120)
* refactor: Refactored src/run.py

* refactor: Cleaned capture script

* feat: Updated dockerfile

* refactor: Moved sensor & controller management to package

* chore: Updated dependencies

* docs: Updated docs

* test: Updated unittests

* fix: Ensures pyroengine is locally built

* fix: Fixed pil resizing

* feat: Added JPEG quality to image saving
