+++ docs/source/core.rst
@@ -1,15 +1,10 @@
 pyroengine.core
 ===============
 
-The models subpackage contains everything to manage the whole Fire Detection process by capturing and saving the image and
-by predicting whether there is a fire on this image
+This module abtracts the sensor management and the detection workflow into controllable units.
 
 
-.. currentmodule:: pyroengine
+.. currentmodule:: pyroengine.core
 
 
-.. autoclass:: Engine
-
-	.. automethod:: clear_cache
-	.. automethod:: predict
-	.. automethod:: heartbeat
+.. autoclass:: SystemController
