+++ docs/source/engine.rst
@@ -1,0 +1,15 @@
+pyroengine.engine
+=================
+
+The engine module contains everything to manage the whole Fire Detection process by capturing and saving the image and
+by predicting whether there is a fire on this image
+
+
+.. currentmodule:: pyroengine.engine
+
+
+.. autoclass:: Engine
+
+	.. automethod:: clear_cache
+	.. automethod:: predict
+	.. automethod:: heartbeat
