+++ docs/source/sensors.rst
@@ -1,0 +1,10 @@
+pyroengine.sensors
+==================
+
+The sensors module provides control over perception sensors connected to the system.
+
+
+.. currentmodule:: pyroengine.sensors
+
+
+.. autoclass:: ReolinkCamera
