+++ src/capture.py
@@ -1,52 +1,54 @@
-# Copyright (C) 2020-2022, Pyronear.
+# Copyright (C) 2022, Pyronear.
 
 # This program is licensed under the Apache License 2.0.
 # See LICENSE or go to <https://www.apache.org/licenses/LICENSE-2.0> for full license details.
 
 
+import argparse
 import json
-import logging
 import os
 import time
-from io import BytesIO
+from pathlib import Path
 
-import requests
-import urllib3
 from dotenv import load_dotenv
-from PIL import Image
 
-urllib3.disable_warnings(urllib3.exceptions.InsecureRequestWarning)
-
-logging.basicConfig(format="%(asctime)s | %(levelname)s: %(message)s", level=logging.INFO, force=True)
+from pyroengine.sensors import ReolinkCamera
 
 
-def capture(ip, CAM_USER, CAM_PWD):
-    url = f"https://{ip}/cgi-bin/api.cgi?cmd=Snap&channel=0&rs=wuuPhkmUCeI9WG7C&user={CAM_USER}&password={CAM_PWD}"
+def main(args):
+    print(args)
 
-    response = requests.get(url, verify=False, timeout=3)
-    return Image.open(BytesIO(response.content))
-
-
-def main():
-    load_dotenv("/home/pi/pyro-engine/.env")
-
+    # .env loading
+    load_dotenv(".env")
     CAM_USER = os.environ.get("CAM_USER")
     CAM_PWD = os.environ.get("CAM_PWD")
+    assert isinstance(CAM_USER, str) and isinstance(CAM_PWD, str)
 
-    with open("/home/pi/pyro-engine/data/credentials.json") as json_file:
+    # Loading camera creds
+    with open(args.creds, "rb") as json_file:
         cameras_credentials = json.load(json_file)
 
-    for ip in cameras_credentials.keys():
-        try:
-            img = capture(ip, CAM_USER, CAM_PWD)
-            file = os.path.join("/home/pi/captured_images", ip, f"{time.strftime('%Y%m%d-%H%M%S')}.jpg")
-            os.makedirs(os.path.split(file)[0], exist_ok=True)
-            img.save(file)
+    # Check if model is available in cache
+    cache = Path(args.cache, "snapshots")
+    cache.mkdir(parents=True, exist_ok=True)
 
-            time.sleep(1)
-        except Exception:
-            logging.warning(f"Unable to get image from camera {ip}")
+    cams = [ReolinkCamera(_ip, CAM_USER, CAM_PWD) for _ip in cameras_credentials]
+
+    imgs = [cam.capture() for cam in cams]
+
+    files = [cache.joinpath(f"{cam.login}_{time.strftime('%Y%m%d-%H%M%S')}.jpg") for cam, img in zip(cams, imgs)]
+
+    for path, img in zip(files, imgs):
+        img.save(path)
 
 
 if __name__ == "__main__":
-    main()
+    parser = argparse.ArgumentParser(
+        description="Camera screenshot script", formatter_class=argparse.ArgumentDefaultsHelpFormatter
+    )
+    # Camera & cache
+    parser.add_argument("--creds", type=str, default="data/credentials.json", help="Camera credentials")
+    parser.add_argument("--cache", type=str, default="./data", help="Cache folder")
+    args = parser.parse_args()
+
+    main(args)
