+++ src/pyproject.toml
@@ -13,5 +13,3 @@ license = "Apache-2.0"
 python = "^3.8"
 pyroengine = "^0.2.0"
 python-dotenv = ">=0.15.0"
-requests = "^2.20.0"
-Pillow = ">=8.4.0"
